feat(.bash_functions): add function to create and activate python virtual environments
The function will first check to see if a virtual environment already exists and activate it. If none exists, it will create and then activate. It will also ask if you want to rename any environments that are not prefixed with a dot.

fix(.bash_aliases): remove pve and pva functions that are no longer needed.

perf(.bashrc): declare .bash_functions file